# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "测试",
        "title": "测试",
        "contact": {},
        "license": {},
        "version": "0.0.1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/team/create": {
            "post": {
                "description": "创建团队",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "1. CreateTeam"
                ],
                "summary": "创建团队接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "团队名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/api/team/dismiss": {
            "post": {
                "description": "查看团队列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "5. ReadListTeam"
                ],
                "summary": "查看团队列表接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页数默认为1",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页数数量默认为20",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/api/team/exit": {
            "post": {
                "description": "退出团队",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "3. ExitTeam"
                ],
                "summary": "退出团队接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "团队id",
                        "name": "tid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/api/team/info": {
            "post": {
                "description": "查看团队信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "6. ReadInfoTeam"
                ],
                "summary": "查看团队信息接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "团队id",
                        "name": "tid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/api/team/join": {
            "post": {
                "description": "加入团队",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "2. JoinTeam"
                ],
                "summary": "加入团队接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "团队id",
                        "name": "tid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/api/team/user_list": {
            "post": {
                "description": "查看团队成员列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "7. ReadUserListTeam"
                ],
                "summary": "查看团队成员列表接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页数默认为1",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页数数量默认为20",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        }
    }
}